feat: add honeypot spam protection to registration
This commit introduces honeypot spam protection to the registration form.

The following changes were made:

- Added the `Spatie\Honeypot\ProtectAgainstSpam` middleware to the global middleware stack.
- Added the `<x-honeypot />` blade component to the registration form.
- Added tests to verify the honeypot field is present and detects spam.
- Moved existing registration tests to `RegisterTest.php`.

This change adds a layer of protection against automated spam submissions on the registration form.
